pull layouts into separate directory
