+++ src/layouts/padPages.test.js
@@ -1,4 +1,4 @@
-import padPages from '../padPages';
+import padPages from './padPages';
 
 const makePage = function () {
   return { element: document.createElement('div') };
