+++ src/main/main.scss
@@ -1,5 +1,8 @@
 $system: -apple-system, BlinkMacSystemFont, "Roboto", sans-serif;
 
 @import '../book/book.scss';
-@import '../Rules/fullBleed.scss';
-@import '../Viewer/viewer.scss';
+@import '../rules/fullBleed.scss';
+@import '../viewer/viewer.scss';
+@import '../viewer/print.scss';
+@import '../layouts/printMarks.scss';
+@import '../layouts/flipLayout.scss';
