+++ src/viewer/viewer.scss
@@ -1,7 +1,3 @@
-@import './Layouts/printMarks.scss';
-@import './Layouts/flipLayout.scss';
-@import './print.scss';
-
 // This file should only include screen-only
 // styles for the viewer that do not persist when
 // printing, but does not yet
